Move all jsPDF calls to PDF wrapper class
